+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Taproot Changelog
 
-## March 2025
+## April 2025
 - `ImuInterface` no longer contains `getTemp` in order to support more inertial measurement sources
 - **Breaking:** All IMU related values are now in radians. (`get[Roll/Pitch/Yaw]()` in rad, `getG[x/y/z]()` in rad/s)
+- Added Butterworth filter coefficient generation.
+    -When constructing a butterworth filter of n order pass in the sample time difference and cutoff frequency in radians/s to obtain a list of coefficients for use in the discrete filter.
+- Added discrete filter object
+    -Takes in a list of natural and forced response coefficients; will compute the next filtered value when .filterData() is called.
+    `.reset()` clears the natural and forced response but keeps the coefficients.
+    `.getLastFilteredValue()` gets the last filtered value.
 
 ## March 2025
 - Added Encoders